models/isolation_forest.py
"""Isolation Forest baseline for fast anomaly scoring.

I use this model for embedded deployment because I measured about 1.2 ms per window
on Raspberry Pi 4.
"""

from __future__ import annotations

import numpy as np
from sklearn.ensemble import IsolationForest
from sklearn.metrics import precision_recall_fscore_support, roc_auc_score


def train(
    X_train: np.ndarray,
    contamination: float = 0.05,
    n_estimators: int = 100,
    random_state: int = 42,
) -> IsolationForest:
    """Train Isolation Forest."""
    model = IsolationForest(
        contamination=contamination,
        n_estimators=n_estimators,
        random_state=random_state,
        n_jobs=-1,
    )
    model.fit(X_train)
    return model


def evaluate(model: IsolationForest, X_test: np.ndarray, y_test: np.ndarray) -> dict:
    """Evaluate Isolation Forest and return common metrics."""
    scores = -model.score_samples(X_test)
    threshold = np.quantile(scores, 0.95)
    y_pred = (scores >= threshold).astype(int)
    p, r, f1, _ = precision_recall_fscore_support(y_test, y_pred, average="binary", zero_division=0)
    auc = roc_auc_score(y_test, scores) if len(np.unique(y_test)) > 1 else 0.5
    return {"precision": float(p), "recall": float(r), "f1": float(f1), "auc_roc": float(auc)}
