requirements.txt
# Core numerical and table processing
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
tqdm==4.66.4

# Machine learning and deep learning models
tensorflow==2.16.1
scikit-learn==1.5.1

# Visualization
matplotlib==3.9.0
seaborn==0.13.2

# Device I/O for Arduino and Raspberry Pi demo
pyserial==3.5

# Utility
joblib==1.4.2
